refactor: moved root_ctx_app_key on agent, storage to prevent circular import
